pyproject.toml
[project]
name = "repl-nix-mouthscan"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "sqlalchemy>=2.0.36",
    "werkzeug>=3.1.3",
    "psycopg2-binary",
    "email-validator",
    "anthropic>=0.42.0",
    "pillow>=11.0.0",
]
